Added logic to fetch data for publicList
Adds the main.js file that contains the client-side logic for this component
Reads the concepts list from the SSR markup, gets the first concept and matches it against the conceptListMapping to get a listId
Fetches the data from myFTClient using the new endpoint getPublicList. However, the endpoint is not ready at the time of this commit.
So, it temporarily adds a fallback with a hardcoded list data
Finally, it hydrates the SSR markup with the listData